+++ src/app/(hospital-finder)/map/page.tsx
@@ -1,0 +1,7 @@
+export default function MapPage() {
+  return (
+    <>
+      <h1>Map Page Goes Here</h1>
+    </>
+  );
+}
