+++ src/app/(landing)/page.tsx
@@ -1,0 +1,11 @@
+export default function HomePage() {
+  return (
+    <>
+      <main>
+        <div>
+          <h1>Landing Page</h1>
+        </div>
+      </main>
+    </>
+  );
+}
